feat: implement geofencing functionality in work location forms and add geofence modal
